refactor(streaming): extract pure seams from hls_service
